refactor(radio-group, radio-group-item)!: Renames radio group components
BREAKING CHANGE: Renames radio group components
- `calcite-radio-group` has been renamed to `calcite-segmented-control`
- `calcite-radio-group-item` has been renamed to `calcite-segmented-control-item`
