--- src/components/radio-group/usage/Basic.md
@@ -1,8 +1,0 @@
-```html
-<calcite-radio-group layout="horizontal" appearance="solid" scale="m" width="auto">
-  <calcite-radio-group-item value="apple" checked>Apple</calcite-radio-group-item>
-  <calcite-radio-group-item value="mango">Mango</calcite-radio-group-item>
-  <calcite-radio-group-item value="tomato">Tomato</calcite-radio-group-item>
-  <calcite-radio-group-item value="banana">Banana</calcite-radio-group-item>
-</calcite-radio-group>
-```
