+++ src/components/segmented-control/usage/Basic.md
@@ -1,0 +1,8 @@
+```html
+<calcite-segmented-control layout="horizontal" appearance="solid" scale="m" width="auto">
+  <calcite-segmented-control-item value="apple" checked>Apple</calcite-segmented-control-item>
+  <calcite-segmented-control-item value="mango">Mango</calcite-segmented-control-item>
+  <calcite-segmented-control-item value="tomato">Tomato</calcite-segmented-control-item>
+  <calcite-segmented-control-item value="banana">Banana</calcite-segmented-control-item>
+</calcite-segmented-control>
+```
