+++ src/demos/form.html
@@ -154,13 +154,13 @@
         <div class="parent">
           <div class="child">
             <calcite-label>
-              Radio group
-              <calcite-radio-group name="calcite-radio-group">
-                <calcite-radio-group-item value="react" checked>React</calcite-radio-group-item>
-                <calcite-radio-group-item value="ember">Ember</calcite-radio-group-item>
-                <calcite-radio-group-item value="angular">Angular</calcite-radio-group-item>
-                <calcite-radio-group-item value="vue">Vue</calcite-radio-group-item>
-              </calcite-radio-group></calcite-label
+              segmented control
+              <calcite-segmented-control name="calcite-segmented-control">
+                <calcite-segmented-control-item value="react" checked>React</calcite-segmented-control-item>
+                <calcite-segmented-control-item value="ember">Ember</calcite-segmented-control-item>
+                <calcite-segmented-control-item value="angular">Angular</calcite-segmented-control-item>
+                <calcite-segmented-control-item value="vue">Vue</calcite-segmented-control-item>
+              </calcite-segmented-control></calcite-label
             >
           </div>
         </div>
