+++ src/index.html
@@ -311,13 +311,6 @@
             </calcite-action>
           </a>
         </div>
-
-        <div>
-          <a href="/demos/radio-group.html">
-            <calcite-action scale="s" text="Radio Group" alignment="start" text-enabled appearance="solid">
-            </calcite-action>
-          </a>
-        </div>
       </div>
 
       <!-- third column -->
@@ -331,6 +324,13 @@
         <div>
           <a href="/demos/scrim.html">
             <calcite-action scale="s" text="Scrim" alignment="start" text-enabled appearance="solid"> </calcite-action>
+          </a>
+        </div>
+
+        <div>
+          <a href="/demos/segmented-control.html">
+            <calcite-action scale="s" text="Segmented Control" alignment="start" text-enabled appearance="solid">
+            </calcite-action>
           </a>
         </div>
