+++ stencil.config.ts
@@ -56,9 +56,9 @@ export const create: () => Config = () => ({
     { components: ["calcite-pick-list", "calcite-pick-list-group", "calcite-pick-list-item"] },
     { components: ["calcite-radio-button"] },
     { components: ["calcite-radio-button-group"] },
-    { components: ["calcite-radio-group", "calcite-radio-group-item"] },
     { components: ["calcite-rating"] },
     { components: ["calcite-scrim"] },
+    { components: ["calcite-segmented-control", "calcite-segmented-control-item"] },
     { components: ["calcite-select", "calcite-option", "calcite-option-group"] },
     { components: ["calcite-shell", "calcite-shell-center-row", "calcite-shell-panel"] },
     { components: ["calcite-slider", "calcite-graph"] },
